Refactor view folders into single files
