--- napari/layers/_base_layer/view/__init__.py
@@ -1,2 +1,0 @@
-from .properties import QtLayerProperties
-from .controls import QtLayerControls
