--- napari/layers/_base_layer/view/controls.py
@@ -1,8 +1,0 @@
-from qtpy.QtWidgets import QFrame
-
-
-class QtLayerControls(QFrame):
-    def __init__(self, layer):
-        super().__init__()
-
-        self.layer = layer
