--- napari/layers/_image_layer/view/__init__.py
@@ -1,2 +1,0 @@
-from .properties import QtImageProperties
-from .controls import QtImageControls
