--- napari/layers/_labels_layer/view/__init__.py
@@ -1,2 +1,0 @@
-from .properties import QtLabelsProperties
-from .controls import QtLabelsControls
