--- napari/layers/_points_layer/view/__init__.py
@@ -1,2 +1,0 @@
-from .properties import QtPointsProperties
-from .controls import QtPointsControls
