--- napari/layers/_shapes_layer/view/__init__.py
@@ -1,2 +1,0 @@
-from .properties import QtShapesProperties
-from .controls import QtShapesControls
