+++ napari/layers/_vectors_layer/view.py
