--- napari/layers/_vectors_layer/view/__init__.py
@@ -1,1 +1,0 @@
-from .properties import QtVectorsProperties
